requirements.txt
jupyter
google-cloud-storage
google-cloud-bigquery
pandas_gbq
python-dotenv
pyarrow
streamlit
plotly